fix: collpase local code into extensions
Previously some of the local code was separated because it had to use hardcoded contracts with different state than the current on-chain versions, which blocked testing of some functionality.

This updates all comments to indicate the appropriate MAINNET/TESTNET values and uses either the TESTNET address or the local version if necessary.

In addition, this removes the need for the script to copy settings between contracts, and removes the pain point of keeping two versions up to date.
