+++ contracts/extensions/ccd011-stacking-payouts.clar
@@ -15,7 +15,7 @@
 
 ;; DATA VARS
 
-;; MAINNET: (define-data-var poolOperator principal 'SPFP0018FJFD82X3KCKZRGJQZWRCV9793QTGE87M)
+;; MAINNET: 'SPFP0018FJFD82X3KCKZRGJQZWRCV9793QTGE87M
 (define-data-var poolOperator principal 'ST1XQXW9JNQ1W4A7PYTN3HCHPEY7SHM6KPA085ES6)
 
 ;; PUBLIC FUNCTIONS
