+++ contracts/proposals/ccip012-bootstrap.clar
@@ -23,16 +23,16 @@
     ))
 
     ;; set 3-of-5 signers
-    ;; MAINNET: (try! (contract-call? .ccd001-direct-execute set-approver 'SP372JVX6EWE2M0XPA84MWZYRRG2M6CAC4VVC12V1 true))
-    ;; MAINNET: (try! (contract-call? .ccd001-direct-execute set-approver 'SP2R0DQYR7XHD161SH2GK49QRP1YSV7HE9JSG7W6G true))
-    ;; MAINNET: (try! (contract-call? .ccd001-direct-execute set-approver 'SPN4Y5QPGQA8882ZXW90ADC2DHYXMSTN8VAR8C3X true))
-    ;; MAINNET: (try! (contract-call? .ccd001-direct-execute set-approver 'SP3YYGCGX1B62CYAH4QX7PQE63YXG7RDTXD8BQHJQ true))
-    ;; MAINNET: (try! (contract-call? .ccd001-direct-execute set-approver 'SP7DGES13508FHRWS1FB0J3SZA326FP6QRMB6JDE true))
-    ;; TESTNET: (try! (contract-call? .ccd001-direct-execute set-approver 'ST3AY0CM7SD9183QZ4Y7S2RGBZX9GQT54MJ6XY0BN true))
-    ;; TESTNET: (try! (contract-call? .ccd001-direct-execute set-approver 'ST2D06VFWWTNCWHVB2FJ9KJ3EB30HFRTHB1A4BSP3 true))
-    ;; TESTNET: (try! (contract-call? .ccd001-direct-execute set-approver 'ST113N3MMPZRMJJRZH6JTHA5CB7TBZH1EH4C22GFV true))
-    ;; TESTNET: (try! (contract-call? .ccd001-direct-execute set-approver 'ST8YRW1THF2XT8E45XXCGYKZH2B70HYH71VC7737 true))
-    ;; TESTNET: (try! (contract-call? .ccd001-direct-execute set-approver 'STX13Q7ZJDSFVDZMQ1PWDFGT4QSBMASRMCYE4NAP true))
+    ;; MAINNET: 'SP372JVX6EWE2M0XPA84MWZYRRG2M6CAC4VVC12V1
+    ;; MAINNET: 'SP2R0DQYR7XHD161SH2GK49QRP1YSV7HE9JSG7W6G
+    ;; MAINNET: 'SPN4Y5QPGQA8882ZXW90ADC2DHYXMSTN8VAR8C3X
+    ;; MAINNET: 'SP3YYGCGX1B62CYAH4QX7PQE63YXG7RDTXD8BQHJQ
+    ;; MAINNET: 'SP7DGES13508FHRWS1FB0J3SZA326FP6QRMB6JDE
+    ;; TESTNET: 'ST3AY0CM7SD9183QZ4Y7S2RGBZX9GQT54MJ6XY0BN
+    ;; TESTNET: 'ST2D06VFWWTNCWHVB2FJ9KJ3EB30HFRTHB1A4BSP3
+    ;; TESTNET: 'ST113N3MMPZRMJJRZH6JTHA5CB7TBZH1EH4C22GFV
+    ;; TESTNET: 'ST8YRW1THF2XT8E45XXCGYKZH2B70HYH71VC7737
+    ;; TESTNET: 'STX13Q7ZJDSFVDZMQ1PWDFGT4QSBMASRMCYE4NAP
     (try! (contract-call? .ccd001-direct-execute set-approver 'ST1SJ3DTE5DN7X54YDH5D64R3BCB6A2AG2ZQ8YPD5 true))
     (try! (contract-call? .ccd001-direct-execute set-approver 'ST2CY5V39NHDPWSXMW9QDT3HC3GD6Q6XX4CFRK9AG true))
     (try! (contract-call? .ccd001-direct-execute set-approver 'ST2JHG361ZXG51QTKY2NQCVBPPRRE2KZB1HR05NNC true))
@@ -43,8 +43,8 @@
     (try! (contract-call? .ccd001-direct-execute set-signals-required u3))
 
     ;; delegate stack the STX in the mining treasuries (up to 50M STX each)
-    ;; MAINNET: (try! (contract-call? .ccd002-treasury-mia-mining delegate-stx u50000000000000 'SP700C57YJFD5RGHK0GN46478WBAM2KG3A4MN2QJ))
-    ;; MAINNET: (try! (contract-call? .ccd002-treasury-nyc-mining delegate-stx u50000000000000 'SP700C57YJFD5RGHK0GN46478WBAM2KG3A4MN2QJ))
+    ;; MAINNET: 'SP700C57YJFD5RGHK0GN46478WBAM2KG3A4MN2QJ
+    ;; MAINNET: 'SP700C57YJFD5RGHK0GN46478WBAM2KG3A4MN2QJ
     (try! (contract-call? .ccd002-treasury-mia-mining delegate-stx u50000000000000 'ST1XQXW9JNQ1W4A7PYTN3HCHPEY7SHM6KPA085ES6))
     (try! (contract-call? .ccd002-treasury-nyc-mining delegate-stx u50000000000000 'ST1XQXW9JNQ1W4A7PYTN3HCHPEY7SHM6KPA085ES6))
